backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21

pillow==12.1.0
opencv-python-headless==4.8.1.78

ultralytics==8.3.252
torch==2.1.2+cpu
torchvision==0.16.2+cpu

numpy==1.26.4

pymongo>=4.6.0
motor>=3.3.0

python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
email-validator>=2.0.0
